requirements.txt
gradio==3.39.0
annoy==1.16.3
flask==2.3.2
torch==2.0.1
tqdm==4.62.3
Pillow==8.2.0
torchvision
pandas
numpy
nltk
transformers
unicodedata2
pickle5


